% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalScClustBench.R
\name{evalScClustBench}
\alias{evalScClustBench}
\title{evalScClustBench}
\usage{
evalScClustBench(sub.kmeans, method = "simlr")
}
\arguments{
\item{sub.kmeans}{An object returned from \emph{scClustBench}}

\item{method}{A method used in \emph{scClustBench}}
}
\value{
A data frame with cluster evaluation for each similarity metric and for each repeats.
}
\description{
Evaluate cluster results from \emph{scClustBench} with evaluation metrics NMI (Normalised Mutual Information), ARI (Adjusted Rand Index), FM (Fowlkes–Mallows) index and Jaccard Index
}
\details{
Evaluate kmeans cluster result from \emph{scClustBench}
}
\examples{
evalScClustBench(sub.simlr, method = "simlr")
evalScClustBench(sub.kmeans, method = "kmeans")

}
