% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmeansEval.R
\name{plotKmeansEval}
\alias{plotKmeansEval}
\title{plotKmeansEval}
\usage{
plotKmeansEval(kmeans.eval, eval.method = c("NMI", "FM", "ARI", "Jaccard"))
}
\arguments{
\item{kmeans.eval}{An object returned from \emph{evalScClustBench}}

\item{eval.method}{Evaluation method to be plotted. Available options are: NMI (default), FM, Jaccard and ARI.}
}
\value{
ggplot object
}
\description{
plot evaluations returned from \emph{evalScClustBench}
}
\examples{


}
